Move client auto tests to new module fabric-client-gametest-api-v1
