--- fabric-api-base/src/testmodClient/resources/fabric-api-base-testmod.client.mixins.json
@@ -1,18 +1,0 @@
-{
-  "required": true,
-  "package": "net.fabricmc.fabric.test.base.client.mixin",
-  "compatibilityLevel": "JAVA_21",
-  "client": [
-    "CyclingButtonWidgetAccessor",
-    "MinecraftClientMixin",
-    "MinecraftDedicatedServerMixin",
-    "ScreenAccessor",
-    "TitleScreenAccessor"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  },
-  "mixins": [
-    "MinecraftServerMixin"
-  ]
-}
