+++ fabric-client-gametest-api-v1/build.gradle
@@ -1,0 +1,6 @@
+version = getSubprojectVersion(project)
+
+moduleDependencies(project, [
+	'fabric-api-base',
+	'fabric-resource-loader-v0'
+])
