+++ fabric-client-gametest-api-v1/src/client/java/net/fabricmc/fabric/impl/gametest/client/ThreadingImpl.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.test.base.client;
+package net.fabricmc.fabric.impl.gametest.client;
 
 import java.util.concurrent.Phaser;
 import java.util.concurrent.Semaphore;
