+++ fabric-client-gametest-api-v1/src/client/resources/fabric.mod.json
@@ -1,0 +1,29 @@
+{
+  "schemaVersion": 1,
+  "id": "fabric-client-gametest-api-v1",
+  "name": "Fabric Client Game Test API (v1)",
+  "version": "${version}",
+  "environment": "client",
+  "license": "Apache-2.0",
+  "icon": "assets/fabric-client-gametest-api-v1/icon.png",
+  "contact": {
+    "homepage": "https://fabricmc.net",
+    "irc": "irc://irc.esper.net:6667/fabric",
+    "issues": "https://github.com/FabricMC/fabric/issues",
+    "sources": "https://github.com/FabricMC/fabric"
+  },
+  "authors": [
+    "FabricMC"
+  ],
+  "depends": {
+    "fabricloader": ">=0.16.9",
+    "fabric-resource-loader-v0": "*"
+  },
+  "description": "Allows registration of client game tests.",
+  "mixins": [
+    "fabric-client-gametest-api-v1.mixins.json"
+  ],
+  "custom": {
+    "fabric-api:module-lifecycle": "experimental"
+  }
+}
