+++ fabric-client-gametest-api-v1/src/testmodClient/resources/fabric.mod.json
@@ -1,0 +1,13 @@
+{
+  "schemaVersion": 1,
+  "id": "fabric-client-gametest-api-v1-testmod",
+  "name": "Fabric Client Game Test API (v1) Test Mod",
+  "version": "1.0.0",
+  "environment": "*",
+  "license": "Apache-2.0",
+  "entrypoints": {
+    "client": [
+      "net.fabricmc.fabric.test.gametest.client.ClientGameTestTestMod"
+    ]
+  }
+}
