+++ gradle.properties
@@ -17,6 +17,7 @@ fabric-biome-api-v1-version=15.0.4
 fabric-block-api-v1-version=1.0.31
 fabric-block-view-api-v2-version=1.0.19
 fabric-blockrenderlayer-v1-version=2.0.7
+fabric-client-gametest-api-v1-version=0.0.1
 fabric-command-api-v1-version=1.2.61
 fabric-command-api-v2-version=2.2.40
 fabric-commands-v0-version=0.2.78
